+++ src/components/Eventos/Eventos.css
@@ -1,110 +1,114 @@
 /* Estilo para o conteúdo principal */
-@import url("https://fonts.googleapis.com/css?family=Roboto&display=swap");
+@import url('https://fonts.googleapis.com/css?family=Roboto&display=swap');
+
+
 
 /*-----------------Header-----------------*/
 
 .header-pos-logon {
-  background-color: #ffffff; /* Cor de fundo do cabeçalho */
-  color: #000000; /* Cor do texto no cabeçalho */
-  height: 50px; /* Altura do cabeçalho fino */
-  display: flex;
-  justify-content: right; /* Centraliza o texto horizontalmente */
-  align-items: center; /* Centraliza o texto verticalmente */
-  border-bottom: 1px solid #e6edff;
+    background-color: #ffffff; /* Cor de fundo do cabeçalho */
+    color: #000000; /* Cor do texto no cabeçalho */
+    height: 50px; /* Altura do cabeçalho fino */
+    display: flex;
+    justify-content: right; /* Centraliza o texto horizontalmente */
+    align-items: center; /* Centraliza o texto verticalmente */
+    border-bottom: 1px solid #E6EDFF;
 }
 
 .header-pos-logon h1 {
-  margin-bottom: 15px;
-  padding: 0;
-  font-size: 16px;
+    margin-bottom: 15px;
+    padding: 0;
+    font-size: 16px;
+    
 }
 
-.imgheader {
-  width: 30px;
-  height: 30px;
-  margin-left: 10px;
-  margin-bottom: 15px;
-  vertical-align: middle;
+.imgheader{
+    width: 30px;
+    height: 30px;
+    margin-left: 10px;
+    margin-bottom: 15px;
+    vertical-align: middle;
 }
 
 /*--------------Menu lateral--------------*/
 
 .menu-lateral {
-  width: 250px;
-  height: 100%;
-  background-color: #fff;
-  position: fixed;
-  bottom: 20px;
-  left: 0;
-  padding: 20px;
-  color: #fff;
-  border-right: 1px solid #e6edff;
+    width: 250px;
+    height: 100%;
+    background-color: #fff;
+    position: fixed;
+    bottom: 20px;
+    left: 0;
+    padding: 20px;
+    color: #fff;
+    border-right: 1px solid #E6EDFF;
 }
 
-.logo-container {
-  padding: 5px;
+.logo-container{
+    padding: 5px;
 }
 
-.logo {
-  width: 100%;
-  height: auto;
+.logo{
+    width: 100%;
+    height: auto;
 }
 .menu-lateral ul {
-  list-style: none;
-  padding: 0;
+    list-style: none;
+    padding: 0;
 }
 
 .menu-lateral ul li {
-  margin-bottom: 20px;
-  margin-left: 20px;
+    margin-bottom: 20px;
+    margin-left: 20px;
 }
 
 .menu-lateral a {
-  text-decoration: none;
-  color: #000000;
-  display: flex; /*flex para manter o ícone e o texto lado a lado */
-  align-items: center;
+    text-decoration: none;
+    color: #000000;
+    display: flex; /*flex para manter o ícone e o texto lado a lado */
+    align-items: center;
 }
 
-.logo-item {
-  width: 24px;
-  height: 24px;
-  margin-right: 8px;
-  vertical-align: middle;
+.logo-item{
+    width: 24px;
+    height: 24px;
+    margin-right: 8px;
+    vertical-align: middle;
 }
 
-.pag-atual {
-  color: #5800d6;
+.pag-atual{
+    color: #5800D6;
 }
 
-.primeiro-menu {
-  position: relative;
+.primeiro-menu{
+position: relative;
 }
-.segundo-menu {
-  position: absolute;
-  bottom: 0;
+.segundo-menu{
+    position: absolute;
+    bottom: 0;
 }
+
 
 /*-----------------Página-----------------*/
-.conteudo {
-  margin-left: 250px;
+.conteudo{
+    margin-left: 250px;
 }
 
-.titulo {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding: 90px 0 40px 100px;
+.titulo{
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding: 90px 0 40px 100px;
 }
 
-.titulo h1 {
-  color: #5800d6;
-  font-size: 35px;
+.titulo h1{
+    color: #5800D6;
+    font-size: 35px;
 }
 
-.titulo h2 {
-  font-size: 15px;
+.titulo h2{
+    font-size: 15px;
 }
 
 .search-bar {
@@ -121,132 +125,58 @@
   border-radius: 7px;
 }
 
-.search-icon {
-  width: 15%;
+.search-icon{
+    width: 15%;
 }
 
-.itens {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 50vh;
-  gap: 10px;
+.itens{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 50vh;
+    gap: 10px;
 }
 
-.calendario {
-  transform: scale(0.8);
+.calendario{
+    transform: scale(0.8);
 }
 
 .itens h1 {
-  font-size: 20px;
+    font-size: 20px;
 }
 
 .botaoPadraoEmpresa {
-  background-color: #4200a1;
-  color: #fff;
-  font-size: 20px;
-  font-weight: 700;
-  border-radius: 6px;
-  border: #4200a1;
-  cursor: pointer;
-  height: 50px;
-  padding: 0 80px;
-}
+    background-color: #4200A1;
+    color: #fff;
+    font-size: 20px;
+    font-weight: 700;
+    border-radius: 6px;
+    border: #4200A1;
+    cursor: pointer;
+    height: 50px;
+    padding: 0 80px;
+  }
 
-.modalEmpresa {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 999;
-}
+  .modalEmpresa{
+    position: fixed;
+    top: 0;
+    right: 0;
+    bottom: 0;
+    left: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 999; 
+  }
 
-.modalContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
+  .boxModal{
+    background-color: white;
+    width: 500px;
+  }
+  .modalContainer{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+  } 
 
-}
-
-.boxModal {
-  background-color: white;
-  width: 800px;
-  border-radius: 12px;
-  height: 85%;
-  overflow-y: scroll;
-}
-
-.boxModal .modalHeader {
-  background-color: #1e1e1e;
-  width: 100%;
-  height: 185px;
-  position: relative;
-}
-
-.boxModal .modalHeader button {
-  background-color: transparent;
-  border: 0;
-  cursor: pointer;
-  color: #fff;
-  position: absolute;
-  right: 20px;
-  top: 20px;
-  font-size: 20px;
-}
-
-.boxModal .modalBody form {
-  display: flex;
-  flex-direction: column;
-  padding: 40px 90px;
-}
-
-.boxModal .modalBody form input,
-textarea {
-  margin-bottom: 20px;
-  padding: 20px;
-  color: #1e1e1e;
-  font-size: 20px;
-  font-weight: 500;
-}
-
-.boxModal .modalBody {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-}
-.boxModal .modalBody img {
-  margin-top: -60px;
-  z-index: 1;
-}
-
-.boxModal .modalBody form {
-  width: 100%;
-}
-
-.boxModal .modalBody form #date {
-  max-width: 200px;
-}
-
-.boxModal .modalBody form label {
-  color: #1e1e1e;
-  font-size: 16px;
-  font-weight: 500;
-  display: block;
-  margin-bottom: 7px;
-}
-
-.boxModal .modalFooter {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 26px;
-}
-
-.boxModal .modalFooter button {
-  border-radius: 12px;
-  background-color: #5800d6;
-}
+
