--- src/components/Eventos/mainContent.css
@@ -1,34 +1,0 @@
-.mainContentEmpresa {
-    margin-left: 250px;
-}
-
-.buttonCreateEvent{
-    max-width: 125px;
-    width: 100%;
-    height: 40px;
-    border: 1px solid #5800D6;
-    background-color: transparent;
-    border-radius: 10px;
-    margin-top: 40px;
-    color: #5800D6;
-    font-weight: bold;
-    font-size: 20px;
-    cursor: pointer;
-}
-
-.titulo{
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding: 90px 0 0px 100px;
-}
-
-.titulo h1{
-    color: #5800D6;
-    font-size: 35px;
-}
-
-.titulo h2{
-    font-size: 15px;
-}
