+++ src/components/TelasUsuario/Maincontent/searchContent.css
@@ -63,6 +63,7 @@
   max-height: 55vh;
   overflow-y: scroll;
   padding-bottom: 100px;
+  width: 100;
 }
 
 .SearchContent .bottom button {
@@ -94,6 +95,7 @@
   justify-content: space-between;
   background: #fff;
   margin-right: 54px;
+  width: 100%;
 }
 
 .SearchContent .bottom .cardUsuario .cardImage {
